usdt_to_btc_test/project_usdt_to_btc_test/requirements.txt
aiodns==2.0.0
aiohttp==3.7.4.post0
appdirs==1.4.3
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
CacheControl==0.12.6
ccxt==1.45.66
certifi==2019.11.28
cffi==1.14.5
chardet==3.0.4
colorama==0.4.3
contextlib2==0.6.0
cryptography==3.4.7
distlib==0.3.0
distro==1.4.0
Django==3.1.7
djangorestframework==3.12.4
html5lib==1.0.1
idna==2.8
ipaddr==2.2.0
lockfile==0.12.2
msgpack==0.6.2
multidict==5.1.0
packaging==20.3
pep517==0.8.2
progress==1.5
pycares==3.1.1
pycparser==2.20
pyparsing==2.4.6
pytoml==0.1.21
pytz==2021.1
requests==2.22.0
retrying==1.3.3
six==1.14.0
sqlparse==0.4.1
typing-extensions==3.7.4.3
urllib3==1.25.8
webencodings==0.5.1
yarl==1.1.0
